Refactor common logic into single method
Reduced some code duplication and saved a few bytes by refactoring some
of the basic +addXXX property trackers into a common function.

Pass in either scrollElement (if specified) or window (if not) to the
tracker during _onScroll.

Changed the trackers incoming param’s name to el from window as a
better indicator of what might be coming in (thought neither is
accurate…)